+++ ui/src/components/Header/Header.tsx
@@ -1,7 +1,7 @@
+import React, { ReactElement } from 'react';
 import { OpenInNew } from '@mui/icons-material';
 import { Box, Typography, Button, AppBar, Toolbar } from '@mui/material';
-import React, { ReactElement } from 'react';
-import { useDDClient } from '../../services/hooks';
+import { useDDClient } from '../../services';
 import { Controller } from './Controller';
 
 export const Header = (): ReactElement => {
