+++ ui/src/components/Header/Menu.tsx
@@ -1,9 +1,9 @@
+import React, { useEffect, useState } from 'react';
 import { MoreVert } from '@mui/icons-material';
 import { IconButton, Menu } from '@mui/material';
-import React, { useEffect, useState } from 'react';
-import { useDDClient } from '../../services/hooks';
+import { useDDClient } from '../../services';
 import { DockerImage } from '../../types';
-import { ConfirmableButton } from '../Feedback/ConfirmableButton';
+import { ConfirmableButton } from '../Feedback';
 import { UpdateDialog } from '../Views';
 
 const ITEM_HEIGHT = 80;
