+++ ui/src/components/Header/index.tsx
@@ -1,2 +1,3 @@
 export * from './Header';
 export * from './Controller';
+export * from './Menu';
