+++ ui/src/components/Views/Logs/LogsPage.tsx
@@ -1,6 +1,6 @@
 import { Box, Card, Typography } from '@mui/material';
 import React, { ReactElement, useEffect, useState } from 'react';
-import { useDDClient, useLocalStack } from '../../../services/hooks';
+import { useDDClient, useLocalStack } from '../../../services';
 
 export const LogsPage = (): ReactElement => {
   const [logs, setLogs] = useState<string[]>([]);
