+++ ui/src/services/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from './api';
 export * from './utils';
+export * from './health';
