+++ ui/src/services/util/format.ts
@@ -1,0 +1,3 @@
+export function Capitalize(str: string): string {
+  return str.charAt(0).toUpperCase() + str.slice(1);
+}
