+++ ui/src/services/util/images.tsx
@@ -2,10 +2,6 @@ import { LATEST_IMAGE } from '../../constants';
 import { DockerImage } from '../../types';
 import { useDDClient } from '../hooks';
 
-export function Capitalize(str: string): string {
-  return str.charAt(0).toUpperCase() + str.slice(1);
-}
-
 export const checkLocalImage = async () => {
   const ddClient = useDDClient();
   const images = await ddClient.docker.listImages() as [DockerImage];
